fix: split renderScene so that locales aren't imported unnecessarily
